+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-### 1.5.1 (Next)
+### 1.5.1
 - More improvements to Packer communicators' settings interfacing with Testinfra provisioner plugin.
 - Improve SSH temporary file cleanup during errors.
 - Improve Testinfra command handling.
 - Improve path validation.
+- Workaround Packer type change for `port` data in new version.
 
 ### 1.5.0
 - Add `transfer_files` parameter.